add check

+++ sources/template/Default/modules/modules_boxes/bm_best_sellers.php
@@ -163,7 +163,7 @@ class bm_best_sellers
 
     $best_sellers = $QBestSellers->fetchAll();
 
-    if (\count($best_sellers) >= MODULE_BOXES_BEST_SELLERS_MIN_DISPLAY && \count($best_sellers) <= MODULE_BOXES_BEST_SELLERS_MAX_DISPLAY) {
+    if (\count($best_sellers) >= (int)MODULE_BOXES_BEST_SELLERS_MIN_DISPLAY && \count($best_sellers) <= (int)MODULE_BOXES_BEST_SELLERS_MAX_DISPLAY) {
       $position = 1;
 
       $bestsellers_list = '<ol class="olBestSellers">';
