+++ sources/template/Default/modules/modules_front_page/fp_new_products.php
@@ -49,7 +49,7 @@ class fp_new_products
     $new_products_category_id = $CLICSHOPPING_Category->getID();
 
     if (CLICSHOPPING::getBaseNameIndex() && !$CLICSHOPPING_Category->getPath()) {
-      if (MODULE_FRONT_PAGE_NEW_PRODUCTS_MAX_DISPLAY != 0) {
+      if ((int)MODULE_FRONT_PAGE_NEW_PRODUCTS_MAX_DISPLAY != 0) {
         if ($CLICSHOPPING_Customer->getCustomersGroupID() != 0) {
           if ((!isset($new_products_category_id)) || ($new_products_category_id == 0)) {
 // Display products no inside categories
