+++ sources/template/Default/modules/modules_products_favorites/ph_products_favorites.php
@@ -45,7 +45,7 @@ class ph_products_favorites
     $CLICSHOPPING_Reviews = Registry::get('Reviews');
 
     if (isset($_GET['Products'], $_GET['Favorites'])) {
-      if (MODULE_PRODUCTS_FAVORITES_MAX_DISPLAY != 0) {
+      if ((int)MODULE_PRODUCTS_FAVORITES_MAX_DISPLAY != 0) {
 
         $products_template = MODULE_PRODUCTS_FAVORITES_TEMPLATE;
